Changed operators argument type from uint_64 to unsigned long long

+++ src/libcsc/libcsc/time/Literals.h
@@ -1,18 +1,18 @@
 #include <time/MyTime.h>
  namespace literals {
-     my_time::Time operator ""_s(uint64_t time) {
+     my_time::Time operator ""_s(unsigned long long time) {
         return my_time::Time(static_cast<time_t>(time));
     }
 
-     my_time::Time operator ""_m(uint64_t time) {
+     my_time::Time operator ""_m(unsigned long long time) {
         return my_time::Time(static_cast<time_t>(time) * my_time::seconds_in_minute);
     }
 
-     my_time::Time operator ""_h(uint64_t time) {
+     my_time::Time operator ""_h(unsigned long long time) {
         return my_time::Time(static_cast<time_t>(time) * my_time::seconds_in_hour);
     }
 
-     my_time::Time operator ""_d(uint64_t time) {
+     my_time::Time operator ""_d(unsigned long long time) {
         return my_time::Time(static_cast<time_t>(time) * my_time::seconds_in_day);
     }
 }
